+++ packages/process/changelog.md
@@ -1,4 +1,7 @@
-#Unreleased
+# Unreleased
+
+# 0.9.0.0
+
 * Remove stderr from the `Process` abstraction so that stdout and stderr must be unified after parsing (potentially as
   `Either` to emulate the previous behaviour).
 * Make default `ProcessOutput` interpreters discard all stderr output.
@@ -11,6 +14,7 @@
   `Either`.
 
 # 0.8.0.0
+
 * Add `ProcessOptions`, replacing the primitive parameters of `Process` interpreters.
 * Add an option for `Process` that determines whether to kill the process after exiting the scope.